Implement server-side caching for playlists

+++ src/services/postgres/PlaylistsService.js
@@ -12,10 +12,11 @@ class PlaylistsService {
    * @param {PlaylistsSongsService} playlistsSongsService
    * @param {CollaborationsService} collaborationsService
    */
-  constructor(playlistsSongsService, collaborationsService) {
+  constructor(playlistsSongsService, collaborationsService, cacheService) {
     this._pool = new Pool();
     this._playlistsSongsService = playlistsSongsService;
     this._collaborationsService = collaborationsService;
+    this._cacheService = cacheService;
   }
 
   /**
@@ -36,6 +37,8 @@ class PlaylistsService {
     if (!result.rowCount) {
       throw new InvariantError('Playlist gagal ditambahkan');
     }
+
+    await this._cacheService.delete(`playlists:${owner}`);
     return result.rows[0].id;
   }
 
@@ -45,17 +48,24 @@ class PlaylistsService {
    * @return {Promise<any[]>}
    */
   async getPlaylists(owner) {
-    const query = {
-      text: `SELECT playlists.id, playlists.name, users.username FROM playlists
-            LEFT JOIN users ON playlists.owner = users.id
-            LEFT JOIN collaborations ON 
-                      collaborations.playlist_id = playlists.id
-            WHERE playlists.owner = $1 OR collaborations.user_id = $1`,
-      values: [owner],
-    };
+    try {
+      const result = await this._cacheService.get(`playlists:${owner}`);
+      return JSON.parse(result);
+    } catch (error) {
+      const query = {
+        text: `SELECT playlists.id, playlists.name, users.username FROM playlists
+              LEFT JOIN users ON playlists.owner = users.id
+              LEFT JOIN collaborations ON 
+                        collaborations.playlist_id = playlists.id
+              WHERE playlists.owner = $1 OR collaborations.user_id = $1`,
+        values: [owner],
+      };
+  
+      const result = await this._pool.query(query);
 
-    const result = await this._pool.query(query);
-    return result.rows;
+      await this._cacheService.set(`playlists:${owner}`, JSON.stringify(result.rows));
+      return result.rows;
+    }
   }
 
   /**
@@ -64,7 +74,7 @@ class PlaylistsService {
    */
   async deletePlaylistById(id) {
     const query = {
-      text: 'DELETE FROM playlists WHERE id = $1',
+      text: 'DELETE FROM playlists WHERE id = $1 RETURNING id, owner',
       values: [id],
     };
 
@@ -73,6 +83,8 @@ class PlaylistsService {
     if (!result.rowCount) {
       throw new NotFoundError('Gagal menghapus playlist. Id tidak ditemukan');
     }
+
+    await this._cacheService.delete(`playlists:${result.rows[0].owner}`);
   }
 
   /**
